sw/python-wrapper/requirements.txt
Flask==2.0.2
lxml==4.8.0
beautifulsoup4==4.11.1
python-dotenv==0.20.0
requests==2.27.1